Revert "Revert "UltimosCambios 07/04""
This reverts commit a2cc7f20a15613c6b6020e531a43528dd5d9698c.

+++ macowins/src/main/java/Grupo4/Saco.java
@@ -6,7 +6,7 @@ public class Saco implements Tipo {
 		this.cantBotones=cantBotones;
 	}
 	
-	public float precioBase(){
+	public double precioBase(){
 		return 300+(10*cantBotones);
 	}
 }
