+++ macowins/src/main/java/Grupo4/Tipo.java
@@ -1,6 +1,6 @@
 package Grupo4;
 
 public interface Tipo {
-	float precioBase();
+	double precioBase();
 	
 }
